Stay inside entity when creating error context
Report the error location within an entity instead of the location of
the entity reference. This is typically more useful and avoids to mess
with ctxt->inputTab.

This should also match behavior of older libxml2 versions more closely
which created a new parser context with a fresh input for general
entities, making the old test ineffective.